Move Node into NodeTree module; minor module renaming
