+++ contextshell/command.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from contextshell.NodePath import NodePath
+from contextshell.path import NodePath
 from contextshell.action import ActionExecutor, parse_argument_tree
