+++ contextshell/exceptions.py
@@ -1,3 +1,3 @@
 
-class NotSupportedError(Exception):
+class NotSupportedError(Exception):  # NOCOVER: uses Exception implementation
     pass
