+++ contextshell/shell.py
