+++ tests/functional/ActionResolvingTests.py
@@ -1,4 +1,4 @@
-from contextshell.NodePath import NodePath
+from contextshell.path import NodePath
 from contextshell.backends.NodeTree import NodeTreeRoot
 from tests.functional.TestExecutor import script_test
 from tests.functional.ShellTestsBase import NodeTreeTestsBase
