+++ tests/functional/FilesystemTests.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from contextshell.action import ActionExecutor
 from tests.functional.ShellTestsBase import TreeRootTestsBase
 from tests.functional.TestExecutor import script_test
-from contextshell.backends.Filesystem import FilesystemRoot
+from contextshell.backends.FilesystemTree import FilesystemRoot
 
 
 class FilesystemTestsBase(TreeRootTestsBase):
