+++ tests/functional/ParserTypesTests.py
@@ -1,6 +1,6 @@
 import unittest
 from contextshell.backends.NodeTree import NodeTreeRoot
-from contextshell.NodePath import NodePath
+from contextshell.path import NodePath
 from tests.functional.ShellTestsBase import NodeTreeTestsBase
 from tests.functional.TestExecutor import script_test
