+++ tests/unit/ArgumentPackerTests.py
@@ -1,6 +1,6 @@
 import unittest
 from contextshell.action import pack_argument_tree, unpack_argument_tree, parse_argument_tree
-from contextshell.NodePath import NodePath as np
+from contextshell.path import NodePath as np
 from collections import OrderedDict
