+++ tests/unit/BuiltinExecutorTests.py
@@ -1,7 +1,7 @@
 import unittest
 
 from tests.unit.Fakes import FakeAction
-from contextshell.NodePath import NodePath
+from contextshell.path import NodePath
 from contextshell.action import BuiltinExecutor
