+++ tests/unit/Fakes.py
@@ -1,4 +1,4 @@
-from contextshell.NodePath import NodePath
+from contextshell.path import NodePath
 from contextshell.action import Action, ActionArgsPack
 from typing import Dict
